+++ src/kubernetes_export/export.py
@@ -1,4 +1,4 @@
-from kubernetes_backup.backup import export_and_backup
+from kubernetes_export.export import export_and_backup
 
 if __name__ == "__main__":
